Rename and reorganize frontend source files
