+++ frontend/src/renderer/src/main.tsx
@@ -2,7 +2,7 @@ import './assets/main.css';
 
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import App from './App';
+import App from './app';
 
 ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
   <React.StrictMode>
